+++ ruby/lib/google/protobuf_native.rb
@@ -1,5 +1,5 @@
 # Protocol Buffers - Google's data interchange format
-# Copyright 2021 Google Inc.  All rights reserved.
+# Copyright 2023 Google Inc.  All rights reserved.
 # https://developers.google.com/protocol-buffers/
 #
 # Redistribution and use in source and binary forms, with or without
@@ -28,6 +28,19 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+# require mixins before we hook them into the java & c code
+require 'google/protobuf/message_exts'
+
+# We define these before requiring the platform-specific modules.
+# That way the module init can grab references to these.
+module Google
+  module Protobuf
+    class Error < StandardError; end
+    class ParseError < Error; end
+    class TypeError < ::TypeError; end
+  end
+end
+
 if RUBY_PLATFORM == "java"
   require 'json'
   require 'google/protobuf_java'
@@ -37,7 +50,30 @@ else
   rescue LoadError
     require 'google/protobuf_c'
   end
+
 end
 
 require 'google/protobuf/descriptor_dsl'
 require 'google/protobuf/repeated_field'
+
+module Google
+  module Protobuf
+
+    def self.encode(msg, options = {})
+      msg.to_proto(options)
+    end
+
+    def self.encode_json(msg, options = {})
+      msg.to_json(options)
+    end
+
+    def self.decode(klass, proto, options = {})
+      klass.decode(proto, options)
+    end
+
+    def self.decode_json(klass, json, options = {})
+      klass.decode_json(json, options)
+    end
+
+  end
+end
